+++ packages/playground/src/index.tsx
@@ -25,13 +25,50 @@
  */
 import React, { FC } from 'react'
 import { render } from 'react-dom'
-import { ComponentsProvider } from '@looker/components'
-import { FieldsDemo } from './Form/FieldsDemo'
+import {
+  ComponentsProvider,
+  FieldRangeSlider,
+  FieldSlider,
+  Grid,
+} from '@looker/components'
 
 const App: FC = () => {
   return (
     <ComponentsProvider>
-      <FieldsDemo />
+      <Grid columns={2}>
+        <FieldSlider
+          detail="detail"
+          description="description"
+          label="Label"
+          min={10}
+          max={20}
+          value={15}
+        />
+        <FieldSlider
+          detail="detail"
+          description="description"
+          label="Label"
+          min={10}
+          max={20}
+          value={15}
+        />
+      </Grid>
+      <Grid columns={2} gap="xxxlarge" mx="20px">
+        <FieldRangeSlider
+          detail="detail"
+          description="description"
+          label="Label"
+          min={300}
+          max={500}
+        />
+        <FieldRangeSlider
+          detail="detail"
+          description="description"
+          label="Label"
+          min={5}
+          max={50}
+        />
+      </Grid>
     </ComponentsProvider>
   )
 }